Refactor users page

+++ app/Http/Controllers/UserController.php
@@ -24,7 +24,7 @@ class UserController extends Controller
             "SPLIT_PART(users.name,' ', -1)" :
             "SUBSTRING_INDEX(users.name, ' ', -1)";
 
-        $user = User::query()->where('name', 'ILIKE', '%' . $search . '%')->orderByRaw($order)->paginate(20);
+        $user = User::query()->where('name', 'ILIKE', '%' . $search . '%')->orderByRaw($order)->get();
 
         if (!empty($request->input('response')) && $request->input('response')) {
             return response()->json($user);
@@ -43,6 +43,10 @@ class UserController extends Controller
         ], [
             'user_id.*' => 'The selected user is invalid or does not exist.'
         ]);
+
+        if (User::query()->find($request->input('user_id'))->admin) {
+            return response()->json(['success' => false, 'message' => 'You cannot delete an admin account.'],403);
+        }
 
         User::query()->where('id', '=', $request->input('user_id'))->where('admin', '=', 'false')->delete();
